♻️ Restructure project following Next.js best practices
- Move secrets to .env.local (JWT_SECRET, HASH_SALT)
- Reorganize directories: utils → lib/, middlewares → middleware/
- Rename css/ → styles/ for consistency
- Convert initDatabase.js → initDatabase.ts
- Configure TypeScript path aliases (@/*)
- Update all imports to use new structure
- Fix duplicate justify-center in AccountItem.tsx
