--- src/pages/api/utils/config.ts
@@ -1,2 +1,0 @@
-export const hashSalt = 10;
-export const jwtSecret = '4a806b0f-eea2-4c0a-b08b-d1543a500a35';
